Move tests for EventEmitter to unit tests
